Roles in iam instead of Resource
1. This is a bit of future proofing. Right now there is a Role resource
https://cloud.google.com/iam/reference/rest/v1/roles/queryGrantableRoles#Role
That defines a role, but can only be listed as roles are only
curated right now. In the future this will be a fully fledged
resource with users able to define custom roles, and I anticipate them
needing a resource class of their own. Hence wrapping the enumeration
of curated roles in this resource (and needing a .name)

2. The same role can be used across multiple resources, and because
of custom roles there's no sense in adding any "sanity check"
code to the mapping between resources and roles. We'll let the IAM
API take care of this for us, and so just move IAM Curated Roles
into the central iam module.